fix(script70): fix IE's script70 error
Whenever we scroll Readium tries to find the next visible media
overlay element to play.
The problem arises when the user changed the page and had scrolled
before. Some funny combination of things triggers the scroll handler to
run again and because it has a stale reference of the iframe's document
IE gives out with the `script70` error message.
This commit unbinds the scroll handler before that can even happen,
i.e., when we are about to open a new page and rebinds it once its done.

May fix #162.