Add "last position" bookmark in reflowable_view
The bookmark is updated every time the pagination changes (in onPaginationChanged).
If there's no deferred page request, at the end of updatePagination,
openPage is called with the bookmark's CFI so whenever viewport is resized or
a font is changed or loaded and the pagination changes, the reader stays on the
same page.
In ReaderView, after a viewport resize, the reader doesn't try to jump to the
bookmark saved at the start of the resize, as it is now handled directly by the
ReflowableView.