Save last page request
Whenever openPage() is called in ReflowView, the pageRequest is saved.
After content is resized (font has changed or viewport has been resized),
we go back to the saved pageRequest.
If the user goes to next or previous page, the saved pageRequest is ditched
and we save the current position in a new pageRequest.