Fix the ResizeSensor by forcing the content document body of a reflowable doc to be "positioned"
The third-party ResizeSensor module wants to set its target element's `position` style to be `relative` as it is key to its functionality.

Unfortunately the condition that checked for this no longer applies to our usage's target element being the 'body'.
It used to check the target element's computed style for position, see if it’s `static` and then change it to `relative`.
As of this patch it no longer does the check in the same way:
https://github.com/marcj/css-element-queries/pull/172

Now it checks the ‘offsetParent’ property of the resize sensor container element, if the offsetParent is not the target element then it applies the ‘relative’ value.
From what I’ve seen by inspecting things with Chrome, looking at the live DOM of the content documents, even if the body has ‘relative’ or not, it by default is the ‘offsetParent’ of the resize sensor element. This means that after this merge the 'position: relative’ that’s needed is no longer applied.