Build scripts into package.json via new method
The scripts declared in CSON syntax are now exclusively in a single file.
The rest of the declarations are now taken from the values of package.json, essentially now "baked-in".

The way it works is by the scripts being compiled from CSON to JSON and the `json` CLI tool merges it into package.json.