Ensure MO player stays paused internally
Internal functionality to MO player and Audio Player is added to "hold"
the playback loop when the player is supposed to be paused.

It's been observed that under certain sitations and specific usages the
pauseMediaOverlay function doesn't consistently keep the playback
paused; MO playback resumes unexpectedly in certain reading system
actions, for example flipping to a new chapter.